pyproject.toml
[build-system]
requires = ['flit_core >=3.2,<4']
build-backend = 'flit_core.buildapi'

[project]
name = "aiohutils"
authors = [{ name = "5j9", email = "5j9@users.noreply.github.com" }]
license = { text = "GNU General Public License v3 (GPLv3)" }
description = "A small library to for aiohttp projects."
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3.10",
]
requires-python = ">=3.12"
dependencies = [
    'aiohttp[speedups]',
    'html5lib',                 # pd.read_html requires html5lib
    'pandas[pyarrow] >= 2.2.1',
]
dynamic = ["version"]

[project.optional-dependencies]
test = ["pytest", "pytest-asyncio", "python-decouple", "pydantic"]


[dependency-groups]
dev = ["aiohutils[test]"]

[tool.uv.sources]
aiohutils = { workspace = true }


[project.readme]
file = "README.rst"
content-type = "text/x-rst"

[project.urls]
Homepage = "https://github.com/5j9/aiohutils"

[tool.ruff]
line-length = 79
format.quote-style = 'single'
lint.isort.combine-as-imports = true
lint.extend-select = [
    'W605', # invalid-escape-sequence
    'FA',   # flake8-future-annotations
    'I',    # isort
    'UP',   # pyupgrade
    'RUF',  # Ruff-specific rules (RUF)
]
lint.ignore = [
    'E721',   # Do not compare types, use `isinstance()`
    'RUF001', # ambiguous-unicode-character-string	
    'RUF002', # ambiguous-unicode-character-docstring
    'RUF003', # ambiguous-unicode-character-comment
    'RUF012', # mutable-class-default
]

[tool.pytest.ini_options]
addopts = '--quiet --tb=short'

[tool.pyright]
typeCheckingMode = 'standard'
reportDeprecated = "warning"
reportPropertyTypeMismatch = "warning"
reportUnnecessaryCast = "warning"
reportUnnecessaryContains = "warning"
reportUnnecessaryIsInstance = "warning"
reportUnnecessaryTypeIgnoreComment = true
reportInvalidStringEscapeSequence = false
reportConstantRedefinition = 'error'
reportTypeCommentUsage = 'warning'
reportUnnecessaryComparison = 'warning'
